fix,refactor: double click on touchpad, scroll without triggering click
