+++ app_pojavlauncher/src/main/java/net/kdt/pojavlaunch/customcontrols/mouse/TapDetector.java
@@ -1,4 +1,4 @@
-package net.kdt.pojavlaunch;
+package net.kdt.pojavlaunch.customcontrols.mouse;
 
 import android.view.MotionEvent;
 
@@ -6,6 +6,8 @@ import static android.view.MotionEvent.ACTION_DOWN;
 import static android.view.MotionEvent.ACTION_POINTER_DOWN;
 import static android.view.MotionEvent.ACTION_POINTER_UP;
 import static android.view.MotionEvent.ACTION_UP;
+
+import net.kdt.pojavlaunch.Tools;
 
 /**
  * Class aiming at better detecting X-tap events regardless of the POINTERS
